Upgrade Bot API (#42)
* upgrade to latest telegram bot api (v5)
* update int64 for user id

+++ models/user_captcha.go
@@ -6,7 +6,7 @@ import "github.com/jinzhu/gorm"
 type UserCaptcha struct {
 	gorm.Model
 
-	UserID    int    `gorm:"index"`
+	UserID    int64  `gorm:"index"`
 	Code      string `gorm:"size:5"`
 	MessageID int
 }
